+++ lib/view/launch/home/state.dart
@@ -15,7 +15,7 @@ abstract class _State extends WidgetState with TickerProviderStateMixin {
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance!.addPostFrameCallback((timeStamp) {
+    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
       // _scrollController.addListener(() {
       //   print('scrolling');
       // });