+++ lib/view/launch/recent_search/main.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: sort_child_properties_last
+
 import 'package:flutter/material.dart';
 
 import 'package:lidea/provider.dart';