+++ lib/view/search/result/bar.dart
@@ -9,15 +9,15 @@ mixin _Bar on _State {
         children: [
           Expanded(
             child: GestureDetector(
-              child: _formField(),
               onTap: onSuggest,
+              child: _formField(),
             ),
           ),
           WidgetButton(
             padding: const EdgeInsets.only(left: 15),
-            child: const WidgetLabel(icon: LideaIcon.home),
             show: args!.hasParam,
             onPressed: param?.currentState!.maybePop,
+            child: const WidgetLabel(icon: LideaIcon.home),
           ),
         ],
       ),